fix(snapshot): isolate Linux mount operations
